// src/models/legacy/actions/Actions.ts
export type ActionType = 'request' | 'response';
// ...
export interface Action {
  /**
   * The type of the action. 
   * The `request` will only process HTTP request data. The `response` has both request and response data available.
   */
  type: ActionType;
  /**
   * Action name.
   */
  name?: string;
  /**
   * Whether the action is enabled.
   */
  enabled?: boolean;
  /**
   * Action priority. The higher number the higher priority of execution.
   */
  priority: number;
  /**
   * Action configuration
   */
  config?: ActionConfiguration;
  /**
   * Whether or not the action is executed synchronously to request / response
   */
  sync?: boolean;
  /**
   * Whether or not the request should fail when the action fails.
   */
  failOnError?: boolean;
  /**
   * Options passed to the UI.
   */
  view?: ActionViewConfiguration;
}

/**
 * Convenience type that gathers all configurations in one type.
 * @deprecated
 */
export type ActionConfiguration = SetCookieConfig | SetVariableConfig | DeleteCookieConfig;

/**
 * @deprecated
 */
export interface SetCookieConfig {
  /**
   * Name of the cookie
   */
  name: string;
  /**
   * Source of the cookie value
   */
  source: DataSourceConfiguration;
  /**
   * When set it uses request URL instead of defined URL in the action
   */
  useRequestUrl?: boolean;
  /**
   * An URL associated with the cookie
   */
  url?: string;
  /**
   * The cookie expiration time
   */
  expires?: string;
  /**
   * Whether the cookie is host only
   */
  hostOnly?: boolean;
  /**
   * Whether the cookie is HTTP only
   */
  httpOnly?: boolean;
  /**
   * Whether the cookie is HTTPS only
   */
  secure?: boolean;
  /**
   * Whether the cookie is a session cookie
   */
  session?: boolean;
}

/**
 * @deprecated
 */
export interface SetVariableConfig {
  /**
   * Name of the variable to set
   */
  name: string;
  /**
   * Source of the cookie value
   */
  source: DataSourceConfiguration;
}

/**
 * @deprecated
 */
export interface DeleteCookieConfig {
  /**
   * When set it uses request URL instead of defined URL in the action.
   */
  useRequestUrl?: boolean;
  /**
   * An URL associated with the cookie.
   * Only used when `useRequestUrl` is not `true`.
   */
  url?: string;
  /**
   * When set it removes all cookies associated wit the URL.
   */
  removeAll?: boolean;
  /**
   * Name of the cookie to remove.
   */
  name?: string;
}
// ...
export interface ActionViewConfiguration {
  /**
   * Whether the action is "opened" in the editor UI.
   */
  opened?: boolean;
}